--- app/src/main/java/org/ole/planet/myplanet/model/Notifications.kt
@@ -1,3 +1,0 @@
-package org.ole.planet.myplanet.model
-
-data class Notifications(var icon: Int, var text: String)
